+++ randomizer/src/components/Button/Button.tsx
@@ -1,24 +1,16 @@
-import React, { useState } from 'react';
+import React from 'react';
 import useUserListContext from '../../contexts/UsersList';
+import { IUserProps } from '../../interfaces';
 import './index.scss';
-
-interface userProps {
-  lastName: string;
-  firstName: string;
-  age: number;
-  gender: string;
-  occupation: string;
-}
 
 const Button: React.FC = () => {
   const userList = useUserListContext();
-  const [, setList] = useState<{ user: userProps; country: string }[]>([]);
 
   function getRandomCountry(arr: string[]) {
     return arr[Math.floor(Math.random() * arr.length)];
   }
 
-  function getRandomUser(arr: userProps[]) {
+  function getRandomUser(arr: IUserProps[]) {
     return arr[Math.floor(Math.random() * arr.length)];
   }
 
@@ -34,8 +26,6 @@ const Button: React.FC = () => {
     const randomUser = getRandomUser(users);
     const randomCountry = getRandomCountry(countries);
     userList.changeUserList({ ...randomUser, country: randomCountry });
-
-    setList((prev) => [...prev, { user: randomUser, country: randomCountry }]);
   }
 
   return (
