+++ randomizer/src/components/Input/Input.tsx
@@ -4,8 +4,8 @@ import useUserListContext from '../../contexts/UsersList';
 import './index.scss';
 
 const Input: React.FC<{ data: string }> = () => {
-  const str = useUserListContext().lastStr;
-  return <input className="app__input" type="text" value={str} disabled />;
+  const { lastStr } = useUserListContext();
+  return <input className="app__input" type="text" value={lastStr} disabled />;
 };
 
 export default Input;
