+++ randomizer/src/contexts/UsersList/Context.ts
@@ -1,21 +1,5 @@
 import React from 'react';
-
-export interface IUser {
-  lastName: string;
-  firstName: string;
-  age: number;
-  gender: string;
-  occupation: string;
-  country: string;
-}
-
-export interface IUserListContext {
-  userList: IUser[];
-  lastStr: string;
-  listStr: string[];
-  counter: number;
-  changeUserList: (user: IUser) => void;
-}
+import { IUserListContext } from '../../interfaces';
 
 const defaultUserList = {
   userList: [],
